refactor(rust): Split protocol.rs into focused domain modules
Split 1,358-line protocol.rs into domain-focused submodules:
- protocol/hook.rs (485 lines): Hook types, builders, 13 tests
- protocol/service.rs (700 lines): Service types, 24 tests
- protocol/mcp.rs (29 lines): MCP types
- protocol/mod.rs (90 lines): Re-exports + protocol-level types

Benefits:
- Focused modules: 23-700 lines vs 1,358-line monolith
- Test locality: Tests live with code they test
- Better git diffs: Hook changes don't touch service types
- API unchanged: All re-exports maintained for backward compatibility

Verification:
- cargo build: clean compilation
- cargo test: all 48 tests passing
- cargo clippy: no warnings
