+++ rust/exomonad-shared/src/protocol/mcp.rs
@@ -1,0 +1,28 @@
+//! MCP (Model Context Protocol) types.
+
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
+
+/// Tool definition for MCP discovery (must match Haskell ToolDefinition).
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ToolDefinition {
+    pub name: String,
+    pub description: String,
+    #[serde(rename = "inputSchema")]
+    pub input_schema: Value,
+}
+
+/// MCP error response.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct McpError {
+    /// Error code.
+    pub code: i32,
+    /// Error message.
+    pub message: String,
+    /// Structured details for debugging.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub details: Option<Value>,
+    /// Actionable guidance to fix the issue.
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub suggestion: Option<String>,
+}
